Module ash_starter::application::render_context

Re-exports

pub use self::frame::Frame;

pub use self::render_target::RenderTarget;

Modules

frame

render_target

Structs

RenderContext
The RenderContext is responsible for requesting an image from the swapchain, picking the corresponding Frame instance, and dispatching to the RenderTarget.

Enums

SwapchainState