Module ash_starter::graphics::draw2d::layer

A layer is a collection of vertices which are all rasterized with the same texture.

Layers are inherently ordered by Draw2D. Layers are drawn from lowest to highest, so layer 5 will be drawn above layer 2.

Structs

Layer

LayerHandle
A layer handle is a unique reference to a layer. Handles can be used to efficiently get a reference to a layer while rendering.

StackedLayers
Stacked Layers is a datastructure which owns all currently available layers. It is responsible for maintaining the render-order for layers when they are added or removed.