Trait ash_starter::graphics::vulkan::buffer::Buffer

pub trait Buffer {
    unsafe fn raw(&self) -> Buffer;

    unsafe fn memory(&self) -> DeviceMemory;

    fn size_in_bytes(&self) -> u64;
}

Required methods

unsafe fn raw(&self) -> Buffer

The raw vulkan handle for the buffer. Should not be copied because implementations are allowed to invalidate the raw buffer value.

Unsafe because it is the responsibility of the caller to ensure that the handle will live for the duration of it's usage. (by checking the specific implementation)

unsafe fn memory(&self) -> DeviceMemory

The raw vulkan handle to the underlying buffer memory. Should not be copied because implementations are allowed to invalidate the raw buffer value.

Unsafe because it is the responsibility of the caller to ensure that the handle will live for the duration of it's usage. (by checking the specific implementation)

fn size_in_bytes(&self) -> u64

The size of the underlying gpu memory in bytes.

Loading content...