Module ash_starter::graphics::vulkan::device::physical_device

Functions for picking a physical device with the features required by this application.

Functions

check_required_extensions
Fetch a vector of all missing device extensions based on the required extensions.

is_device_suitable
Return true when the device is suitable for this application.

pick_physical_device
Pick a physical device based on suitability criteria.

required_device_extensions
Return the set of required device extensions for this application

required_device_features
Return the set of required device features for this application.