Struct ash_starter::rendering::device::Device

pub struct Device {
    pub physical_device: PhysicalDevice,
    pub logical_device: Device,
    pub graphics_queue: Arc<Queue>,
    pub present_queue: Arc<Queue>,
    pub window_surface: Arc<dyn WindowSurface>,
    pub instance: Arc<Instance>,
}

This struct holds all device-specific resources, the physical device and logical device for interacting with it, and the associated queues.

Fields
physical_device: PhysicalDevicelogical_device: Devicegraphics_queue: Arc<Queue>present_queue: Arc<Queue>window_surface: Arc<dyn WindowSurface>

A reference to the window surface used to create this device

instance: Arc<Instance>

The Vulkan library instance used to create this device

Implementations

impl Device

pub fn new(window_surface: Arc<dyn WindowSurface>) -> Result<Arc<Device>>

Create a new device based on this application's required features and properties.

pub fn name_vulkan_object<Name, Handle>(
&self,
name: Name,
object_type: ObjectType,
handle: &Handle
) -> Result<()> where
Handle: Handle + Copy,
Name: Into<String>,

Give a debug name for a vulkan object owned by this device.

Whatever name is provided here will show up in the debug logs if there are any issues detected by the validation layers.

Example

device.name_vulkan_object(
    "Graphics Queue",
    vk::ObjectType::QUEUE,
    &device.graphics_queue()
)?;